requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Userbot (БЕЗ TgCrypto для Windows)
pyrogram==2.0.106

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Environment
python-dotenv==1.0.0

# Utils
requests==2.31.0
aiohttp==3.9.1